refactor: Collectors
