+++ packages/core/src/definitions/types/base-collector-metadata.ts
@@ -2,5 +2,5 @@ import { CollectorMetadata } from '../../resolvers/collector/collector-metadata'
 
 export type BaseCollectorMetadata = Omit<
   CollectorMetadata,
-  'metadata' | 'type'
+  'metadata' | 'type' | 'classInstance'
 >;
