+++ packages/core/src/resolvers/collector/resolved-collector-infos.ts
@@ -1,7 +1,10 @@
-import { CollectorMetadata } from '../collector-metadata';
+import { ModuleRef } from '@nestjs/core';
+
+import { CollectorMetadata } from './collector-metadata';
 
 export interface ResolvedCollectorInfos {
   instance: InstanceType<any>;
   methodName: string;
+  moduleRef: ModuleRef;
   collectors: CollectorMetadata[];
 }
