+++ packages/core/src/resolvers/event/event.resolver.ts
@@ -4,7 +4,7 @@ import { ClientEvents } from 'discord.js';
 import { ExecutionContext } from '../../definitions/interfaces/execution-context';
 import { ReflectMetadataProvider } from '../../providers/reflect-metadata.provider';
 import { DiscordClientService } from '../../services/discord-client.service';
-import { CollectorResolver } from '../collector/use-collectors/collector.resolver';
+import { CollectorResolver } from '../collector/collector.resolver';
 import { FilterResolver } from '../filter/filter.resolver';
 import { GuardResolver } from '../guard/guard.resolver';
 import { MethodResolveOptions } from '../interfaces/method-resolve-options';
