+++ packages/sample/interaction-collector/src/app.module.ts
@@ -1,7 +1,7 @@
 import { DiscordModule } from '@discord-nestjs/core';
 import { Module } from '@nestjs/common';
 import { ConfigModule, ConfigService } from '@nestjs/config';
-import { Intents, Message } from 'discord.js';
+import { Intents } from 'discord.js';
 
 import { BotModule } from './bot/bot.module';
 
@@ -23,8 +23,6 @@ import { BotModule } from './bot/bot.module';
         registerCommandOptions: [
           {
             forGuild: configService.get('GUILD_ID_WITH_COMMANDS'),
-            allowFactory: (message: Message) =>
-              !message.author.bot && message.content === '!deploy',
             removeCommandsBefore: true,
           },
         ],
