+++ packages/sample/interaction-collector/src/bot/bot.module.ts
@@ -2,10 +2,9 @@ import { DiscordModule } from '@discord-nestjs/core';
 import { Module } from '@nestjs/common';
 
 import { PlayCommand } from './command/play.command';
-import { PostInteractionCollector } from './post-interaction-collector';
 
 @Module({
   imports: [DiscordModule.forFeature()],
-  providers: [PlayCommand, PostInteractionCollector],
+  providers: [PlayCommand],
 })
 export class BotModule {}
