+++ packages/sample/message-collector/src/bot/bot.module.ts
@@ -1,11 +1,8 @@
-import { DiscordModule } from '@discord-nestjs/core';
 import { Module } from '@nestjs/common';
 
 import { BotGateway } from './bot.gateway';
-import { QuizMessageCollector } from './quiz-message-collector';
 
 @Module({
-  imports: [DiscordModule.forFeature()],
-  providers: [BotGateway, QuizMessageCollector],
+  providers: [BotGateway],
 })
 export class BotModule {}
