+++ packages/sample/reaction-collector/src/bot/bot.module.ts
@@ -1,11 +1,8 @@
-import { DiscordModule } from '@discord-nestjs/core';
 import { Module } from '@nestjs/common';
 
-import { AppreciatedReactionCollector } from './appreciated-reaction-collector';
 import { BotGateway } from './bot.gateway';
 
 @Module({
-  imports: [DiscordModule.forFeature()],
-  providers: [BotGateway, AppreciatedReactionCollector],
+  providers: [BotGateway],
 })
 export class BotModule {}
