+++ packages/sample/sub-command/src/app.module.ts
@@ -1,7 +1,7 @@
 import { DiscordModule } from '@discord-nestjs/core';
 import { Module } from '@nestjs/common';
 import { ConfigModule, ConfigService } from '@nestjs/config';
-import { Intents, Message } from 'discord.js';
+import { Intents } from 'discord.js';
 
 import { BotModule } from './bot/bot.module';
